Unify shader code and add new fp32 shader

+++ shaders/mandelbrot_fp32.frag
@@ -2,7 +2,7 @@
  * spnda, Copyright (c) 2020
  * base shader fragment shader
  */
-#version 460 core
+#version 330 core
 
 uniform int height;
 uniform int width;
@@ -16,17 +16,17 @@ layout(location = 0) out vec4 colour;
 
 int MAX_ITERATIONS = 100;
 
-double julia(dvec2 z, dvec2 c) {
+float julia(vec2 z, vec2 c) {
 	for (int i = 0; i < MAX_ITERATIONS; i++) {
-		dvec2 r = z * z;
+		vec2 r = z * z;
 		if (r.x + r.y > 4.0) return i;
-		z = dvec2(r.x - r.y + c.x, 2.0 * z.x * z.y + c.y);
+		z = vec2(r.x - r.y + c.x, 2.0 * z.x * z.y + c.y);
 	}
 	return MAX_ITERATIONS;
 }
 
-vec3 rgb(double ratio) {
-    double normalized = ratio * 256 * 6;
+vec3 rgb(float ratio) {
+    float normalized = ratio * 256 * 6;
     int x = int(mod(normalized, 256.0));
     int red = 0, grn = 0, blu = 0;
     switch (int(normalized / 256)) {
@@ -40,32 +40,32 @@ vec3 rgb(double ratio) {
 	return vec3(red, grn, blu) / 255.0;
 }
 
-vec3 gradient(double colour) {
+vec3 gradient(float colour) {
 	vec2 pt = vec2(colour);
-	dvec3 color1 = dvec3(1.0, 0.55, 0.0) * colour;
-	dvec3 color2 = dvec3(0.226, 0.0, 0.615) * colour;
-	double mixValue = distance(pt, vec2(0,1));
+	vec3 color1 = vec3(1.0, 0.55, 0.0) * colour;
+	vec3 color2 = vec3(0.226, 0.0, 0.615) * colour;
+	float mixValue = distance(pt, vec2(0,1));
 	return vec3(mix(color1, color2, mixValue));
 }
 
-vec3 backgroundGradient(double colour) {
+vec3 backgroundGradient(float colour) {
 	vec2 pt = gl_FragCoord.xy/vec2(width, height).xy;
-	dvec3 color1 = dvec3(1.0, 0.55, 0.0) * colour;
-	dvec3 color2 = dvec3(0.226, 0.0, 0.615) * colour;
-	double mixValue = distance(pt, vec2(0,1));
+	vec3 color1 = vec3(1.0, 0.55, 0.0) * colour;
+	vec3 color2 = vec3(0.226, 0.0, 0.615) * colour;
+	float mixValue = distance(pt, vec2(0,1));
 	return vec3(mix(color1, color2, mixValue));
 }
 
-double translateCoordinates(double val, int maxV, int minV) {
+float translateCoordinates(float val, int maxV, int minV) {
     return ((val - (minV)) / (maxV - (minV))) * 4.0 - 2.0;
 }
 
 void main(void) {
-	double x = translateCoordinates(int(gl_FragCoord.x), width, 0);
-	double y = translateCoordinates(int(gl_FragCoord.y), height, 0);
+	float x = translateCoordinates(int(gl_FragCoord.x), width, 0);
+	float y = translateCoordinates(int(gl_FragCoord.y), height, 0);
 	x = x * zoom + positionX;
 	y = y * zoom + positionY;
-	double fractal = julia(dvec2(x, y), dvec2(x, y)) / MAX_ITERATIONS;
+	float fractal = julia(vec2(x, y), vec2(x, y)) / MAX_ITERATIONS;
 	vec3 colourOut = rgb(fractal);
 	colour = vec4(colourOut, 1.0f);
 }
